bee gees uninitialised variable carry

+++ vm/regs.c
@@ -12,22 +12,23 @@
 
 #include "vm.h"
 
-int			valid_reg_num(uint8_t reg_num)
+int			valid_reg_num(uint32_t reg_num)
 {
 	return (reg_num > 0 && reg_num < REG_NUMBER);
 }
 
-int			is_reg_num_invalid(uint8_t reg_num)
+int			is_reg_num_invalid(uint32_t reg_num)
 {
 	return (reg_num == 0 || reg_num >= REG_NUMBER);
 }
 
-uint32_t	get_reg_val(t_process *process, uint8_t reg_num)
+uint32_t	get_reg_val(t_process *process, uint32_t reg_num)
 {
+	//ft_printf("get_reg_val %i\n", reg_num);
 	return (process->regs[reg_num]);
 }
 
-void		set_reg_val(t_process *process, uint8_t reg_num, uint32_t new_val)
+void		set_reg_val(t_process *process, uint32_t reg_num, uint32_t new_val)
 {
 	if (is_reg_num_invalid(reg_num))
 		return ;
